+++ apis/vara-eth/src/api/index.ts
@@ -7,3 +7,4 @@ export {
   InjectedTxReceipt,
   TransactionPurgedReason,
 } from './injected/index.js';
+export * from './types/index.js';